+++ src/Fields/Table.php
@@ -99,4 +99,15 @@ class Table extends Field_Base {
 	public function set_table_options( array $options ): void {
 		$this->table_options = $options;
 	}
+
+	/**
+	 * The sanitize callback for this field.
+	 *
+	 * @param mixed $value The value to save.
+	 *
+	 * @return mixed
+	 */
+	public function default_sanitize_callback( mixed $value ): mixed {
+		return $value;
+	}
 }