+++ src/Fields/Text.php
@@ -113,7 +113,7 @@ class Text extends Field_Base {
 	 *
 	 * @return string
 	 */
-	public function sanitize_callback( mixed $value ): string {
+	public function default_sanitize_callback( mixed $value ): string {
 		// bail if value is null.
 		if ( is_null( $value ) ) {
 			return '';